[WIP] Move `arrow_schema_to_tiledb_schema` to `arrow_adapter.h`
