+++ apis/python/src/tiledbsoma/_dataframe.py
@@ -233,17 +233,14 @@ class DataFrame(pts.SOMADataFrame):
         ptrs_domain = []
         if domain is not None:
             for dim_name, dom in zip(index_column_names, domain):
-            # for sch, dom in zip(schema, domain):
                 dtype = schema[schema.get_field_index(dim_name)].type
                 arr = pa.array(dom, type=dtype)
-                print(arr, dtype)
                 c_domain = ffi.new("struct ArrowArray*")
                 ptr_domain = int(ffi.cast("uintptr_t", c_domain))
                 gc.collect()
                 arr._export_to_c(ptr_domain)
                 ptrs_domain.append(ptr_domain)
         
-        print(ptrs_domain)
         
         # CHECK DOMAIN AND INDEX COL NAME SIZE
